requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.27
pydantic>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
pyyaml>=6.0
pytest>=7.0
pytest-cov>=4.0
black>=24.0
flake8>=7.0
mypy>=1.0
isort>=5.13
tenacity>=8.0
